safari-534.53-branch/LayoutTests/sputnik/Conformance/11_Expressions/11.2_Left_Hand_Side_Expressions/11.2.4_Argument_Lists/S11.2.4_A1.3_T1-expected.txt

CONSOLE MESSAGE: line 79: SyntaxError: Parse error
S11.2.4_A1.3_T1

PASS Expected parsing failure

TEST COMPLETE
